@@ -50,7 +50,7 @@
     "@codeyourfuture/eslint-config-standard": "^3.0.3",
     "@testing-library/cypress": "^8.0.2",
     "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^13.0.0",
     "axe-core": "^4.4.0",
     "babel-loader": "^8.2.4",
     "concurrently": "^7.0.0",
@@ -72,8 +72,8 @@
     "msw": "^0.39.2",
     "nodemon": "^2.0.14",
     "prop-types": "^15.8.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.2",
     "rimraf": "^3.0.2",
     "style-loader": "^3.3.1",
