@@ -1,10 +1,10 @@
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 import App from "./App";
 
-ReactDOM.render((
+createRoot(document.getElementById("root")).render((
 	<BrowserRouter>
 		<App />
 	</BrowserRouter>
-), document.getElementById("root"));
+));
